refactor flux-feed, ocr-feed and total
change package structure
split wizard-logic if necessary
refactor method names and naming in general
