+++ src/modules/reward/feed/ocr/service.ts
@@ -1,11 +1,13 @@
-import * as OcrAggregator from '../../../../artifacts/OffchainAggregatorBilling.json';
-import * as wizardText from '../../../../resources/wizard.json';
+import * as OcrAggregator from '../../../../../artifacts/OffchainAggregatorBilling.json';
+import * as wizardText from '../../../../../resources/wizard.json';
 import { BigNumber, Contract, providers, utils } from 'ethers';
-import { BillingSet, FeedRewardStatus } from '../../../model/feed_reward_status';
-import { AddressInfo } from '../../../model/address_info';
-import { Helper } from '../../../helper/help';
-import { TrxInfo } from '../../../model/trx_info';
+import { BillingSet, FeedRewardStatus } from '../../../../model/feed_reward_status';
+import { AddressInfo } from '../../../../model/address_info';
+import { Helper } from '../../../../helper/help';
+import { TrxInfo } from '../../../../model/trx_info';
 import _ from 'lodash';
+
+const GAS_USAGE: BigNumber = BigNumber.from(231255);
 
 export class OcrFeedRewardService {
   async _getTrxInfo(provider: providers.BaseProvider, transactionHash: string): Promise<TrxInfo> {
@@ -18,8 +20,8 @@ export class OcrFeedRewardService {
     let message: string = '';
     if (!feedStatus.rewardData.linkWeiPerTransmission.eq(billingSet.linkWeiPerTransmission)) {
       message += wizardText.ocr_feed_wizard.replies.reward_per_transmission_change.format(
-        Helper.getLinkValueWithDefinedDecimals(feedStatus.rewardData.linkWeiPerTransmission, 4),
-        Helper.getLinkValueWithDefinedDecimals(billingSet.linkWeiPerTransmission, 4)
+        Helper.parseLinkWeiToLink(feedStatus.rewardData.linkWeiPerTransmission, 4),
+        Helper.parseLinkWeiToLink(billingSet.linkWeiPerTransmission, 4)
       );
     }
     if (!feedStatus.rewardData.linkPerEth.eq(billingSet.linkPerEth)) {
@@ -39,6 +41,19 @@ export class OcrFeedRewardService {
       const newCurrentGasPrice: BigNumber = currentGasPrice.add(gasBonus);
       return gasUsed.mul(newCurrentGasPrice).mul(billingSet.linkPerEth).add(billingSet.linkWeiPerTransmission);
     }
+  }
+
+  _getAverageTransmitterReward(
+    currentFeedStatus: Map<string, FeedRewardStatus<BillingSet>>,
+    gasPrice: BigNumber
+  ): BigNumber {
+    let totalTransmitterRewards: BigNumber = BigNumber.from(0);
+    for (const feedStatus of currentFeedStatus.values()) {
+      totalTransmitterRewards = totalTransmitterRewards.add(
+        this._getNewTransmitterReward(feedStatus.rewardData, gasPrice, GAS_USAGE)
+      );
+    }
+    return totalTransmitterRewards.div(currentFeedStatus.size);
   }
 
   _isTransmitterRewardChanged(
@@ -86,7 +101,6 @@ export class OcrFeedRewardService {
     for (const feedStatus of currentFeedStatus.values()) {
       const currentBillingSetList: number[] = await feedStatus.contract.getBilling();
       const currentBillingSet: BillingSet = this._createBillingSet(currentBillingSetList);
-      console.log(currentBillingSet.linkPerEth.toString());
       feedStatus.rewardData = currentBillingSet;
     }
   }
